Sync Win: work around breakage in IJ 1.46d
The behavior of ImageJ1 changed with version 1.46d: now the StackWindow
will not open an ImagePlus as hyperstack when it was opened as hyperstack
before. You now have to set the openAsHyperStack flag explicitly.

Since that flag was introduced in version 1.39g (2 Nov 2007), Sync Windows
now requires at least that version. Sorry.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>