Add a script to synchronize multiple remote repositories
The idea is to call this script with the URLs of the remote repositories
to synchronize. For performance, you can specify the git:// URL followed
by "=<push URL>":

	git-synchronizer.sh \
		git://fiji.sc/fiji.git=fiji.sc:/srv/git/fiji.git \
		git://github.com/fiji/fiji=github.com:fiji/fiji

The idea is that a remote is added for each URL whose name is constructed
from the URL. Then, the remote branches are compared before and after
fetching to determine which branches want to be deleted, updated, or
pushed for the first time. Conflicting information is printed as an error
and none of the conflicting updates are performed.

An additional step makes sure that fast-forwarding updates are
synchronized between the remotes (again, conflicts are marked as errors).

The goal is to let Jenna Jenkins performs the job in regular intervals.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>