ImageJ launcher: properly escape --edit arguments
On Windows, just to be different, the directory separator is a backslash.
Unfortunately, the backslash happens to be the escape character in strings
almost everywhere, including the ImageJ macro language. So we need to
escape the backslashes before passing them to ImageJ.

This fixes the issue that

	ImageJ-win32 --edit path/to/file.java

would not open the file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>