Add a pom.xml file for the submodules
This configuration file uses Maven's profile mechanism to activate
submodules based on the presence of the pom.xml file.

That way, submodules that are checked out are built from source while
compile artifacts are downloaded from the Maven repository for those
submodules which are not checked out. Just the way it should be.

This also fixes our problem when the snapshot versions of imglib2 advanced
in a way incompatible with the imagej2 revision committed in fiji.git; now
we just require all the developers who check out imagej2 to also check out
imglib and they will end up with revisions that match.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>