ij-launcher: add getClassPath() method
Technically, we would not need to build our list of URLs since the
URLClassLoader keeps track of them.

However, for JavaDoc, we need to be able to build the classpath to
pass to the JavaDoc main class without adding more components to the
Java classpath (lest incompatible doclets are found, breaking JavaDoc).

So we will introduce a flag to stop adding components to the Java
classpath, but still allow getClassPath() to return the full class
path option to be passed to JavaDoc.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>