ij-launcher: add -freeze-classloader
Sometimes, we want to keep building the class path so it can be passed
to the main() method using the -pass-classpath option, but we do not
necessarily want that classpath to be used for loading the main class.

This option allows exactly that: it stops adding to the current
ClassLoaderPlus' classpath, but let's the logic handle -jarpath and
similar options for retrieval via getClassPath().

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>