ImageJ launcher (Windows XP): try harder to find the launcher's DOS name
When the name lacks the .exe extension, XP's GetShortPathName() does not
work properly. Earlier, this developer could not reproduce the problem
because the tests were performed on a Windows 7 box (so evidently, it is
an XP-only problem).

While at it, fix the issue when the launcher is actually not in the
current directory but instead in the PATH.

Reported by Sebastian Rhode as issue 345.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>