ij-launcher: -jarpath and -ijjarpath really only want .jar files
A recent change in the ImageJ launcher broke javadoc-all because it tried
to include all directories in all the sources as class paths. Of course,
that way it found files we wanted to exclude explicitly.

Instead, force -jarpath to perform what it was meant to do and
incidentally fix javadoc-all.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>