ImageJ launcher: heed jvm.cfg setting when falling back to a smaller heap
When the memory setting in jvm.cfg cannot be accomodated, we want to
fall back to trying again with a smaller heap. However, the previous
code did not know how to parse that memory setting, so it used "0"
instead, failing.

With this change, we are now looking harder for the memory setting.

Reported by Olivier Bardot as issue 356.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>