Teach MiniMaven to retrieve a parent pom.xml if necessary
In the scijava context, we want to have a central parent pom.xml in which
all the version numbers and the repositories are listed, to avoid
duplication and conflicts. Since that parent pom.xml should not need to be
present in the local working directory, MiniMaven needs to be able to
detect that situation and fetch the file.

Requested by Curtis Rueden.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>