TrackMate_: WIP: trying to fix the visible track indices bug
We save all the tracks and the indices of the track that are selected to
be visible.

Unfortunately, when we load the data, we do not load the track one by
one, but we rebuild the graph edge by edge. Then, we pass the newly
created graph to the model, and the tracks are regenerated by calling
the connected components.

So there is no reason for the saved track indices to match the indices
of the regenerated tracks, which is what happen as soon as we manually
add a spot somewhere before saving.

I see no elegant way to solve this now. Will spend some time under the
thinking hat.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>