MiniMaven: findPOM() requires explicit downloadAutomatically parameter
In some cases it might be desirable to be able to find a POM without
defaulting to downloading it, even when MiniMaven's downloadAutomatically
field was set to true.

For example, in the next commit we will add the ability to fall back to
reading the parent directory's pom.xml if the parent pom was not yet
parsed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>