Updater (cmdline): let the UserInfo determine the username
If the user name is not set in db.xml.gz, it might be set in .netrc. The
UserInfo will determine that, no need to annoy the uploading user with an
unnecessary question.

This is required for Jenkins to be able to upload non-interactively
whatever upstream ImageJ throws at us.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>