Update javac from icedtea6
This is really painful; OpenJDK6's javac does not match the sources
from the official jdk6 Mercurial repository, ImgLib does not compile with
our former javac, SPIM_Registration does not compile with the javac we
just merged from IcedTea6, it's all a big mess.

Conclusion: do not use the latest and greatest and buggiest javac
features. Unless you do not have to fix anything yourself but can rely on
idiot Johannes to try to fix it for everybody relying on default javac.

When he should take care of the ImageJ launcher in ImageJ2 instead. Cool
thing, to be able on such an idiot doing all the nitty gritty grunt work.

After this commit, SPIM_Registration still does not compile, but that
idiot will figure out from the diff what needs to be done to make it work
again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>