Add imglib2.jar as dependency for loci_tools.jar
This is important to ensure that ImgLib2 builds before Bio-Formats does.
At the moment, it is not strictly necessary, but Bio-Formats develop
branch HEAD now depends on imglib2 core, and next time we update Fiji's
Bio-Formats submodule pointer, we will need this dependency in place.

So we might as well add it now.