Build.sh: exclusively use bin/ImageJ.sh
Earlier, we tried very hard to use the platform-specific launchers for
building. But that does not really make much sense since we already have a
shell script version of the ImageJ launcher (which can call Fiji Build),
and Build.sh being a shell script to begin with, we require a shell
anyway.

So: remove a lot of the complexity by only special-casing jars/fake.jar
(which cannot be helped...).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>