Refresh Javas: imitate (unfortunate) ImageJ1 behavior faithfully
ImageJ 1.* does not require classes to be instances of PlugIn or
PlugInFilter to run them as plugins; it falls back to instantiate the
classes with the default (no-arguments) constructor.

Refresh Javas tried to encourage better programming by falling back to
executing the static main() method.

To stay as bugwarts-compatible as possible, we still try to run main() but
fall back to trying the default constructor.

Reported by Christophe Leterrier as bug 405.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>