Fiji Build: use ImageJ.sh in the fall-back rule
Even if we will not use the fall-back rule for the launcher, it is safer
to use the shell script rather than relying on the launcher to be
up-to-date.

The original plan was to teach MiniMaven to special-case the ImageJ
launcher build from IJ2's sources and fall back to use the Fiji Updater.

However, MiniMaven is really not meant to replace Maven but instead to
provide the minimal functionality we need to build Fiji without adding the
burden of a complete Maven "lets-download-half-the-internet" build.

To this end, it is enough to rely on the ImageJ launcher to be uploaded to
the Fiji Updater and not build the launcher from source in the default
Fiji Build anymore.

If need be, we can always add a very simple script later to use Maven and
copy things in-place for those who need to build/debug the launcher.

This will allow us to rely on the updater to get the ImageJ launcher
when modules/imagej2/ was not checked out.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>