Manual Tracking: re-add support for tracking in Z
By default, the tracking is done in the time dimenension. However,
if the image has only a single frame and more than one z slice, the
plugin will do the tracking in the z dimension. To reduce the
introduction of complexity, the access to information about the
tracking dimension (current position, maximum) is factored out in
separate methods.