download-launchers: convert ij-launcher.jar to Java5, too
Retrotranslator cannot (yet) JIT Java6 to Java5 code. In Fiji, we
consequently compiled everything for Java5, therefore, and so we also need
a Java5-compatible ij-launcher.jar. In the near future, however, we will
need to teach Retrotranslator that new trick, because ImageJ2 is destined to
be compiled always to conform with Java6.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>