Build.sh: work around MSys bug
Apparently, MSys' PATH mangling (to replace colons by semicolons as
mandated by Windows' DOS heritage)does not like PATH to begin with a
relative path. So let's use an absolute one.

The symptom was this error message when trying to download the launcher:
'sh' is not recognized as an internal or external command, operable
program or batch file.

Reported by Jean-Yves Tinevez.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>