Fiji Updater: protect against Java5's lack of File.canExecute()
Java5 does not have that method, and Retrotranslator does not provide it,
either. So for the benefit of the poor souls still stuck with Java5
(thanks, Redmont, thanks very much), fail gracefully when we cannot
determine whether a file is executable or not.

This is not really relevant, either, since the 'executable' flag needs to
be set only when uploading the first time, and people who try to do that
using Java5 cannot be helped anyway.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>