Add a script to build Retrotranslator
For now, this is a pretty manual process. Should we ever need it to be
more automatic, IMHO it would be best to Mavenize it anyway.

That would be a bit involved, though, as Retrotranslator requires several
JDKs to be available. At the moment, modules/Retrotranslator/3rdparty/
contains these projects to allow a successful build here:

	jdk1.3.1_20/
	j2sdk1.4.2_17/
	jdk1.5.0_16/
	mx4j-3.0.2/
	xalan-j_2_7_1/

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>