Use StageLabel if no Plane elements are found
This commit refactors stage position parsing logic into a common helper
method. It is more careful to access Plane metadata only if that plane's
metadata is present. Failing that, it tries to fall back to the
StageLabel metadata if no plane position metadata is found.

The StageLabel stores the X/Y/Z stage position for an entire Image,
and is a good fallback in case the Plane Position[XYZ] is absent.