Updater: special-case the checksum of Maven .properties
The Properties class writes out text files where the second line is a
comment recording the time when things were saved. Maven uses that
mechanism for pom.properties (and if it is a NAR project like
ij-launcher, also nar.properties).

This interfers with our Updaters' checksumming which assumes that the
files _contents_ remain unchanged between builds.

So let's special-case .properties files and simply skip all
commented-out lines when checksumming.

For backwards-compatibility with the Fiji Updater and with earlier
versions of the ImageJ Updater (we identify whether a file is locally
modified by comparing the checksum with a list of known checksums), we
only do that when the .properties file is newer than the cut-off date
(June 15th 2012).

[This commit was cherry-picked from ImageJ2, to help transition to the
ImageJ Updater in the near future.]

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>