download-launchers.sh: get .exe files with icons provided by Jenkins
It is a major hassle to attach the correct icon to a .exe file, especially
when we're not linking the executable but want to edit an existing one.

There are quite a number of tools to edit the icon and other resources in
existing .exe, but all of them are written for Windows (hence do not run
on this developer's machine, not even using WINE which seems to have
gotten slower and buggier over the past few years). Furthermore, most of
them do not offer a command-line interface and if they do, they are closed
source so you cannot really trust them.

Except WinRun4J's rcedit.exe.

The download-launchers.sh script already knows how to run that executable
but that is pretty useless if we can only run it on Windows machines.

Jenkins to the rescue. We have a Jenkins job that builds our launchers
already, so why not use it to change the icon? This is exactly what we do
now: it is a downstream project of ImageJ-launcher called
Windows-Fiji-launcher-with-icons and it publishes exactly two artifacts:
fiji-win32.exe and fiji-win64.exe. With the Fiji icon.

Since that job is already done and since we already rely on Jenkins having
admirably performed the job to compile the launchers, we can just download
those artifacts also on Windows just as well.

This will avoid the problem that yours truly forgot to make sure that
executables _with_ icon are uploaded into the Updater by uploading from a
Windows machine, as noticed by Rainer M. Engel. We no longer need to
upload them from Windows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>