TrackMate_: refactor settings object
It makes more sense to store all feature filters in the settings object,
rather that in the model which owns the settings object. That way, the
settings object contains all the parameters the user can set, and the
model will contain all data resulting from analysis.

This change does not affect the XML file format, but for consistency, I
will change it in a future release.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>