javac: fix compile problems with ImgLib
ImgLib really challenges the generic-parsing-fu of any java compiler.
Unfortunately, we did not have access to the source code of any javac that
compiled ImgLib successfully, so we had to provide a couple of nasty, ugly
workarounds.

The solution I found is based on a mixture of the current OpenJDK6 javac
source code, a single method borrowed from the initial OpenJDK7 javac
source code, and the insight that all of our problems were caused by an
incomplete generic type resolution.

I realized that whenever we have compile errors with our javac that we do
not have with Sun's javac of JDK6 Update 24 nor with Eclipse's compiler,
the instantiateMethod() method of com.sun.tools.javac.comp.Infer handles a
list of unresolved generic types (restvars) that is not empty.

A lesson I learnt from Knud Poulsen: if something fails, just try again.
So I copied the code that resolves the generic types and tried again, and
got less compile errors. Wrapping the resolution code into a loop that
repeats as often as new types could be resolved finally fixed my
compilation problems.

As the bulk of the change is an indentation change, it might be nicer to
look at this commit using 'git show -w'.

Now, this might not be the most elegant solution, and it might be too
greedy, resolving too-specific types, but it compiles all of Fiji, and if
people do not like it and do not come up with a better solution
themselves, they can come, find & kiss me.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>