fiji-compat: add a very simple memory profiler
This memory profiler instruments all method entries and exists using
javassist. At each exit, it reports the relative memory usage, the total
memory usage, and the exit point of the current method. Call it like this:

	./fiji --main-class fiji.MemoryProfiler fiji.Main

Since the memory profiling slows down execution dramatically due to the
synchronous output to stderr, you may want to limit the classes to be
instrumented by setting the environment variable MEMORY_PROFILE_ONLY to a
space-delimited list of classes.

If you want to instrument any class handled by IJPatcher, you need to use
the slightly more complicated command line:

	./fiji -Dpatch.ij1=false \
		--cp jars/javassist.jar --cp jars/fiji-compat.jar \
		--cp jars/ij.jar --main-class fiji.MemoryProfiler -- \
		ij.ImageJ

I wrote this class to profile the sudden resource hunger of Fiji Build and
figured that this class might come in handy in the future, or for others,
or both.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>