Fiji Build: never ask MiniMaven for a dependency without groupId or version
The only such occasion was when we asked to build an artifact from a
submodule, since MiniMaven then had to search for the correct POM in the
tree.

But then, we only support a handful of Mavenized submodules, and almost
all of them have the same groupId and version throughout all of their
pom.xml files, including the root one. So we can use that one.

The only notable exception is ImgLib1 -- which unfortunately is pretty
resistant to taking its well-deserved retirement. So we have to
special-case that. Sigh.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>