Special-case the first ImageJ Updater run
We want to yse the ImageJ Updater which recently has been forked off of
the Fiji Updater with the idea to move the component to ImageJ2. But
that means that we want to access the ImageJ Updater from within Fiji
without having its sources anymore.

So the idea is to instantiate the Updater itself from a known-good
version on the ImageJ Update site and let it install the ImageJ
Updater. After that, we use the ImageJ Updater happily ever after.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>