MiniMaven: special-case the Fiji Updater
Since the original Fiji Updater checksummed the contents of the .jar
files byte for byte, output from Maven was considered different with
every new build. For this reason, the ImageJ Updater special-cases
a couple of file types and skips e.g. the timestamps in .properties
files.

Since we would like the original Fiji Updater to update itself, at
least, and since it verifies the checksum of the downloaded file,
we need to make sure that the file Fiji_Updater.jar does not contain
such fiddly files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>