Bead registration: work around an unwelcome behavior of GenericDialog
This developer can only hope that at least one person benefits from the
funny behavior of GenericDialog in that it does not record any options
when a DialogListener was added. Because it is quite unintuitive (one
might even consider it a nasty side effect) that the recording is turned
off after the *first* (and *only* the first) DialogListener has been sent
a null event, but before the showDialog() method returns.

To work around this, we now call getNextString() and friends in the
DialogListener, when that infamous null event occurs, store the result in
a field of Bead_Registration, and then return it from the thread that
called showDialog().

Noticed by Stephan Preibisch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>