TrackMate_: WIP: change design to factory
Let's face it: the way we are offering to extend TrackMate is unelegant.
Change it to factory design: the plugin is given some factories that
provides it with detectors, trackers, etc...
A developer willing to plug its own algorithm would just have to provide
a factiry with its own algoirthms.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>