Fiji Updater: try harder to run ImageJ's updater
The Updater is really the heart of Fiji (and now of ImageJ2, too), so we
really, really need to keep it functional.

So when somebody messed up (most likely yours truly), we need to try much
harder to update. There might be a NoClassDefFoundError for example, if
the ImageJ2 components that are available locally are not matching.

Since we already have one known-okay version of the updater which we use
to update the updater itself using a remote class loader, let's fall back
to running the updater from said remote class loader to run the (old, but
working) updater from there if the local updater does not want to start.

Of course, this change needs to trickle down to the Fiji users who update
their installations, because the Fiji Updater no longer updates itself
before doing anything else anymore...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>