TrackMate_: changed the deiusgn for the detectors
Again, changed the plugin design, taking inspiration from some walks on
Saint-Palais beach.

The goal is to concile, even clumsily, the ability to run the plugin
from a script and a GUI in a simple manner, and to extend this plugin,
for a script, and for the GUI.

This design decouples a bit the complexity of the GUI from the script
part, and concentrate clumsiness in a single class: the detector
provider. It manages the list of the available detectors for the GUI,
and manages save to & load from XML.

Someone only wanted to extend for scripts does not have to deal with all
these, and still use TrackMate with his own detectors.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>