precompiled.sh commit: strip CR characters
Like everybody else, Maven and Java pay a time and effort tribute to
Redmont, using DOS-style line endings in the manifest. To be able
to work with the commit name stored in the manifest, we therefore
need to strip it out again, paying our own time and effort tax to
Redmont.

Thanks, guys.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>