Build.sh: avoid using the fall-back rule
The fall-back rule which used the updater to avoid committing
precompiled .jar files used to look like a good idea, but was not: the
updater serves stable files while we want to work with unstable
versions.

A much better idea was to implement a minimal build system that groks
Maven project configuration files and re-uses the Maven repositories
established by the ImageJDev project.

So let's rely exclusively on the Maven repositories for precompiled .jar
files (called "artifacts" in Maven speak).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>