Add a Javascript to help recover the updater
We had a couple of problems with the updater, mainly because we moved
it to ImageJ2 and made a couple of changes there that did not consider
running outside of an ImageJ2 context.

To help the situation, this Javascript ignores every local class and
runs exclusively from a known-good set of .jar files from the ImageJ
update site.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>