Add a Beanshell script to create a Linux .desktop file
On Linux (many X11 setups, really), the standard way to add an icon to the
desktop is by creating a .desktop file.

Launching an application from such a .desktop file also allows the icon to
be locked to Ubuntu's Unity's launcher.

Unfortunately, the format of .desktop files requires absolute paths, so we
have to generate it on the fly.

Wished-for by Ilan Tal.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>