create-jnlp: we self-host Java3D now
As pointed out by Kees Straatman, the official Webstart'able Java3D .jar
files are no longer signed. The reason is explained here:

	https://forums.oracle.com/forums/thread.jspa?messageID=9882107

Oracle considers it a security risk to maintain the Java3D and JOGL
projects and continue to offer packages signed by Sun.

So we had to sign the .jar files and host them ourselves, which we now do.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>