TrackMate_: completely reqork the track features logic
There is now a kind of caching mechanism: the FeatureModel is notified
when the model changes, and invalidate track features if adequate.
Features marked as invalid are recomputed on the fly if required
thereafter.
Works well to update a view - colors are based on features - after a
manual modification.
Expect bugs and future changes.
The caching is really simple: all features are invalidated for all track
at once, and validity is restored feature by feature, but for all
tracks. More subtle approaches could be investigated where the feature
value of only a few tracks would be invalidated, left the other fine.
But that requires more work on the track feature analyzers: they need to
declare if they can tolerate a local change, or if the smallest change
invalidate the value for all tracks.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>