Bump the SciJava POM to 1.27
This makes the plugin copying more implicit by simply setting
the imagej.app.directory property appropriately (which the
imagej-maven-plugin picks up, as configured by the SciJava POM).

Actually, due to a programming mistake by this developer, we still need to
specify the imagej-maven-plugin's version explicitly for the moment, as
the imagej-maven-plugin as used by the SciJava POM had a problem where it
would not pick up the imagej.app.directory property correctly from the
pom.xml file (but would work correctly when the property was passed via
the command-line).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>