Fix null and duplicate entries in choices list
This reverts a small portion of commit
bb9e78a703d2816a30287f139625770f0e667192.

The problem was that the allChoices list: A) had a null entry;
and B) had a duplicate entry. Null entries are definitely not allowed in
GenericDialogs (raising an NPE with message "cannot add null item to
Choice"), and duplicate entries are at least unnecessary and at worst
problematic. This change addresses both issues, and allows macros to
successfully invoke the Stitching plugins again.