Stitching_: check errors after phase correlation
The ImgLib1 class PhaseCorrelation can log an error during its process()
method, which should be checked after being called. (I double checked,
and it is checked everywhere else it is used; just this one place
ignored it.)

We also now check for a null result during pairwise stitching and
collection stitching in general, and issue error messages if that
happens, to be sure the user is informed that something went wrong.