Change the user-plugins' path to $HOME/.plugins/
In some cases, users want to install plugins even if Fiji is installed
system-wide (and thus not writable by the user).

We support this now: any .jar file in ~/.plugins/ will be added to the
menu structure, too.

Wished for by a person on IRC who did not quite manage to communicate
effectively.

There is one obstacle which prevented us from going forward with user
plugins earlier: What if user plugins want to override system-wide
installed plugins? This is still not possible. But rather than solve a
problem that users might not even have, we punt on that issue for now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>