Fix Plugins>Utilities>Create Desktop Icon
The problem with Java programs and X11 desktops such as GNOME or Unity is
that the Java windows do not have a distinguising WM_CLASS. Hence Unity's
dock's refusal to allow the user to "Lock to Launcher". And hence the
duplicate dock entry when launching from a .desktop.

The trick is to set the StartupWMClass property in the .desktop file
(unfortunately, this file is still required to run Fiji, and even more
unfortunate is the fact that it cannot contain relative paths, so we
cannot ship one ourself).

This fix was a huge wish of Ilan Tal's.

While at it, we also fixed the Categories entry, as pointed out by
desktop-file-validate(1).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>