PerformanceProfiler: add a more convenient method to start profiling
The new way is to call

	if (PerformanceProfiler.startProfiling(null, args)) return;

at the beginning of the main(args) method of the code to profile.

This tells PerformanceProfiler to instrument that main class and load it
into a custom class loader.  When launched again in the instrumented class
loader, startProfiling() will realize that profiling already started and
return false.

Thus, the instrumented main method will be executed and when the original,
uninstrumented main() method continues execution, startProfiling() will
have returned true and the rest will not be executed again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>