Build.sh: make sure that uploading via ssh will work
Recent architecture changes showed the problem (again, we had the
same problem with Fiji Build's missing-precompiled-fallback) with
relying on the update site: it will never be completely up-to-date.

In the case of Ignacio Arganda-Carreras (as reported as bug 575),
ij-updater-ssh was taken from the update site and did not match
ij-updater-core as taken from the Maven repository.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>