StartupMacros: discover and run macros in macros/AutoRun/
The biggest shortcoming of the StartupMacros mechanism (which
it shares with the macro interpreter's Ext.* mechanism as well as
with the HandleExtraFileTypes class) is that there can be only one
file. There is no easy way that you can augment it locally.

Instead of forcing users to have their own, locally-modified
StartupMacros file that will be forever stale, change our StartupMacros
so that it discovers all macros stored as .ijm files in macros/AutoRun/
and runs them, in alphabetical order (for consistency).

Users who need to install macros upon startup (e.g. for keyboard
shortcuts) can always call

	run("Install...", "install=[" + getDirectory("imagej")
		+ "/macros/my-shortcuts.ijm]");

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>