Pay Janelia Farm a time tribute
Some institutes find it funny to block arbitrary ports, probably
due to a lack of understanding what the ports are for. For example,
port 9418 -- well-established git:// port -- is blocked in one of
the institutes one of the Fiji founders works in.

Since it is easier to change oneself than it is to change somebody
else, we simply activated smart HTTP transport and switched .gitmodules
to use it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>