Alt listener: work around Xpra releasing Alt before switching focus
Xpra unfortunately behaves differently than X.org when it comes
to switching windows via Alt+Tab: it first releases the Alt key
and then nothing happens for a couple of milliseconds. However, we
really do not want to show the menu when switching between windows
(also because the opened menu then blocks input from the window we
switched to). Therefore, use the freshly-added class to invoke the
code much, much later and hope that Xpra switched focus by then.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>