IJHacker: eliminate repeated ImageJ from version
Previously, the version string -- as returned by version() and in the
status bar -- showed "(Fiji Is Just) ImageJ/ImageJ" as a holdover from
when it was "Fiji/ImageJ". It made sense at the time but is now no
longer needed.

This change is mostly just an excuse for me to try out the new WebDAV
upload mechanism. ;-)