IJHacker: optionally prevent batch mode from exiting
When running ImageJ in batch mode from Java code, we might not want it
to exit always. Unfortunately, there is no way to do so with plain
ImageJ, hence we add code to the IJHacker to handle the command-line
option '-batch-no-exit' that will prevent ImageJ#main from exiting.

This is very useful for testing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>