Add an integration test for running bare plugins
A bare plugin is a plain .class file in the plugins/ directory. There is
a problem with that, though: placing a .class file into a subdirectory
does not say whether the class is supposed to live in the package
corresponding to the subdirectory's name, or whether the file was just
put there to place the corresponding menu item into a submenu of the
Plugins menu.

We fixed this so many times that this developer got really tired of
it, so let's have an integration test that is run by Maven so that
we can really, really be sure to catch breakages early.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>