Make the report about a changed method signature less confusing
ImageJ users -- for which this handling caters -- usually do not
understand exceptions. Hence we tried to give a bit more intuitive
information about the case when a method signature changed --
probably due to conflicting class path entries. However, we still
started the report with an exception which confuses the heck out
of non-programmers.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>