Segmentation Editor: Prevent modification of interpolation result.
Probably the only simple way to make this even more secure (the result
can still be modified during the final copying phase) would be to
create a copy of the ROI array.  This would mean a substantial memory
overhead though.

Previously if a slice with an empty ROI was selected during
interpolation, it would stay empty even if it should have received
some ROI.