Volume Viewer: move classes to non-default package
The reason is that classes in the default package take silent precedence
over classes made available via wildcard imports, which has caused
confusion on multiple occasions.

For example, the statement "import java.awt.*;" does *not* import the
java.awt.TextField class when a packageless "TextField" class is also on
the classpath.

See this thread for further rationale and discussion:
https://groups.google.com/d/msg/fiji-devel/6Ua05iccSVo/5ZLrUO-FhmUJ