Remove the ImageJA submodule
Since our switch to Maven, it is very clear which ImageJ 1.x version
we use: the one defined in the pom-scijava version that we use as
our parent POM.

Since the Maven configuration can refer to a completely different ImageJ
1.x version than is checked out in the submodule (and Maven would rather
download the specified version than build the submodule), let's just
remove it altogether to avoid confusion.

For documentation purposes, and for ease of use with existing setups,
let's keep the information in .gitmodules and the support to include
building ImageJA together with Fiji in modules/pom.xml, though.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>