TrackMate_: WIP: redo the track model
The previous model becomes extremely slow against manual punctual
modifications when there is a lot of tracks. This commit and the
following ones aim at replacing the old model by a better one, that
listen to these changes, and update the model accordingly, but in an
incremental manner.
The goal is to have a performance when modifying tracks that depends
only weakly on the number of tracks.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>