TrackMate_: solve a severe performance issue with TrackScheme
TrackScheme was taking ages to render a medium to large model (1000s of
tracks 100000s of spots). This was not so much due to the size of the
model than bad programming:

When updating the style of all cells, we were updating all the edges
between a model.beginUpdate() / model.endUpdate() transaction block.
This is effective.
But we forgot to do so when updating the vertices. So each vertex update
was generating a seris of event. Using a transaction block for this as
well solves the performance issue.

Rendering a model took previously 800s. Now it takes 20s.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>